# 4668 performance regression in unrolling and circuit to instruction