--- a/qiskit/circuit/library/standard_gates/ms.py
+++ b/qiskit/circuit/library/standard_gates/ms.py
@@ -45,5 +45,5 @@ class MSGate(Gate):
         for i in range(self.num_qubits):
             for j in range(i + 1, self.num_qubits):
                 rules += [(RXXGate(self.params[0]), [q[i], q[j]], [])]
-        qc.data = rules
+        qc._data = rules
         self.definition = qc