--- a/qiskit/circuit/library/standard_gates/u2.py
+++ b/qiskit/circuit/library/standard_gates/u2.py
@@ -70,7 +70,7 @@ class U2Gate(Gate):
         q = QuantumRegister(1, 'q')
         qc = QuantumCircuit(q, name=self.name)
         rules = [(U3Gate(pi / 2, self.params[0], self.params[1]), [q[0]], [])]
-        qc.data = rules
+        qc._data = rules
         self.definition = qc
 
     def inverse(self):