--- a/qiskit/converters/ast_to_dag.py
+++ b/qiskit/converters/ast_to_dag.py
@@ -390,7 +390,7 @@ class AstInterpreter:
             op = self._create_op(child_op.name, params=eparams)
             rules.append((op, qparams, []))
         circ = QuantumCircuit(qreg)
-        circ.data = rules
+        circ._data = rules
         return circ
 
     def _create_dag_op(self, name, params, qargs):