--- a/qiskit/extensions/hamiltonian_gate.py
+++ b/qiskit/extensions/hamiltonian_gate.py
@@ -108,7 +108,7 @@ class HamiltonianGate(Gate):
         """Calculate a subcircuit that implements this unitary."""
         q = QuantumRegister(self.num_qubits, 'q')
         qc = QuantumCircuit(q, name=self.name)
-        qc.append(UnitaryGate(self.to_matrix()), qargs=q[:])
+        qc._append(UnitaryGate(self.to_matrix()), q[:], [])
         self.definition = qc
 
     def qasm(self):